[image-manipulator] Fix handling urls from the photo library
